pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skull-netrecon"
version = "1.0.0"
description = "Professional Network Reconnaissance Tool"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "SKULL Security Team", email = "contact@skull-security.io"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: Proprietary",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]
dependencies = [
    "scapy>=2.5.0",
    "python-nmap>=0.7.1",
    "netifaces>=0.11.0",
    "rich>=13.7.0",
    "colorama>=0.4.6",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "jinja2>=3.1.2",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
]

[project.scripts]
skull-netrecon = "skull_netrecon.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["skull_netrecon*"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
